@@ -1,12 +1,14 @@
 {
   "name": "peternguyen.com-react",
-  "version": "0.1.0",
+  "homepage": "http://peternguyen.com",
+  "version": "1.2.0",
   "private": true,
   "dependencies": {
-    "react": "^16.7.0",
-    "react-dom": "^16.7.0",
-    "react-router-dom": "^4.3.1",
-    "react-scripts": "2.1.3"
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-router-dom": "^6.4.0",
+    "react-scripts": "2.1.3",
+    "react-transition-group": "^4.4.5"
   },
   "scripts": {
     "start": "react-scripts start",
