@@ -1,16 +1,17 @@
 import React, { Component } from "react";
+import { Link } from "react-router-dom";
+import Page from './UI/Page.js';
 
 class Contact extends Component {
 	render() {
 		return (
-			<div className="intro centered">
-				<div class="text">
-					<p><a href="mailto:peter7n@yahoo.com">Email</a></p>
-					<p><a href="https://www.linkedin.com/in/peter7n/">LinkedIn</a></p>
-					<p><a href="https://twitter.com/master_ptn">Twitter</a></p>
-					<p><a href="https://github.com/peter7n">GitHub</a></p>
-				</div>
-			</div>
+			<Page className="intro centered">
+				<p><a href="mailto:peter7n@yahoo.com">Email</a></p>
+				<p><Link to="/resume">Resumé</Link></p>
+				<p><a href="https://www.linkedin.com/in/peter7n/">LinkedIn</a></p>
+				<p><a href="https://twitter.com/master_ptn">Twitter</a></p>
+				<p><a href="https://github.com/peter7n">GitHub</a></p>
+			</Page>
 		);
 	}
 }
